Move common test methods to separate utils file
